scripts/flask_backend/requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
numpy>=1.24.0
google-generativeai>=0.3.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
